Redmineのチケットナンバー　 #31820
　不具合状況の説明：物件登録画面に削除ボタンが表示されている。
　修正内容の説明：
物件番号の有無に関わらず、削除ボタンを表示する処理となっている。
そのため、if文を追加し、物件番号が無い場合は表示しない処理となるようソースコードを書き換えました。。

+++ SRC/libArticle.php
@@ -260,9 +260,11 @@ function subArticleEdit()
 			</tr>
 		</table>
 
-		<a href="javascript:fnArticleEditCheck();"><img src="./images/<?php print $btnImage ?>" /></a>　
+		<a href="javascript:fnArticleEditCheck();"><img src="./images/<?php print $btnImage ?>" /></a>
 		<a href="javascript:form.act.value='articleSearch';form.submit();"><img src="./images/btn_return.png" /></a>
-		&nbsp;&nbsp;<a href="javascript:fnArticleDeleteCheck(<?php print $articleNo ?>);"><img src="./images/btn_del.png" /></a>
+		<?php if ($articleNo) { ?>
+			&nbsp;&nbsp;<a href="javascript:fnArticleDeleteCheck(<?php print $articleNo ?>);"><img src="./images/btn_del.png" /></a>
+		<?php } ?>
 	</form>
 <?php
 }
@@ -334,6 +336,8 @@ function subArticleDelete()
 
 	$sql = fnSqlArticleDelete($articleNo);
 	$res = mysqli_query($conn, $sql);
+	// var_dump($sql);
+	// exit;
 
 	$_REQUEST['act'] = 'articleSearch';
 	subArticle();
